+++ RSM/database/migrations/2019_07_20_142221_create_orders_table.php
@@ -15,13 +15,13 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('cid')->nullable();
-            $table->bigInteger('did')->nullable();
+            $table->integer('cid')->nullable();
+            $table->integer('did')->nullable();
             $table->string('description');
             $table->string('Color')->nullable();
             $table->string('PN')->nullable();
             $table->string('UnitMeasurment')->nullable();
-            $table->Integer('quantity');
+            $table->integer('quantity');
             $table->bigInteger('uid');
             $table->string('notes')->nullable();
             $table->timestamps();
